fixed a strange performance issue related to the background image and added a scroll animation helper function
